I'm still develop it
There is only one way for me to motivate myself to work hard: I don’t think about it as hard work. I think about it as part of making myself into who I want to be. Once I’ve made the choice to do something, I try not to think so much about how difficult or frustrating or impossible that might be; I just think about how good it must feel to be that, or how proud I might be to have done that. Make hard look easy. 


I didn't get enough time for  to learner  because I 
think 🤔  , it's easy to learn because  there is big talented support like pluralslight, andela ,